Move SETTLEMENT_INTERVAL constant to model module
