+++ daemon/src/model/cfd.rs
@@ -13,10 +13,10 @@ use crate::model::Timestamp;
 use crate::model::TradingPair;
 use crate::model::TxFeeRate;
 use crate::model::Usd;
+use crate::model::SETTLEMENT_INTERVAL;
 use crate::olivia;
 use crate::olivia::BitMexPriceEventId;
 use crate::payout_curve;
-use crate::SETTLEMENT_INTERVAL;
 use anyhow::bail;
 use anyhow::Context;
 use anyhow::Result;
