+++ daemon/src/projection.rs
@@ -21,8 +21,8 @@ use crate::model::Price;
 use crate::model::Timestamp;
 use crate::model::TradingPair;
 use crate::model::Usd;
+use crate::model::SETTLEMENT_INTERVAL;
 use crate::Order;
-use crate::SETTLEMENT_INTERVAL;
 use anyhow::Context;
 use anyhow::Result;
 use async_trait::async_trait;
